# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: aa@bb.com
  description: Ambulance Waiting List management for Web-In-Cloud system
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
  title: Waiting List Api
  version: 1.0.0
servers:
- description: SwaggerHub Api Auto Mocking
  url: https://virtserver.swaggerhub.com/MichalSevcik/AmbulanceWaitingList/1.0.0
tags:
- description: Secured Admin-only calls
  name: ambulanceAdmins
- description: Operations available to regular developers
  name: ambulanceDevelopers
paths:
  /waiting-list/{ambulanceId}:
    get:
      description: |
        By using ambulanceId you can retrieve waiting list of that particular
        ambulance.
      operationId: getAmbulanceDetails
      parameters:
      - description: pass the ambulanceId of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                response:
                  value:
                    id: dentist-warenova
                    name: Zubná ambulancia Dr. Warenová
                    roomNumber: 356 - 3.posch
                    waitingList:
                    - id: x321ab3
                      name: Jožko Púčik
                      patientId: 460527-jozef-pucik
                      waitingSince: 2038-12-24T10:05:00.000Z
                      estimatedStart: 2038-12-24T10:35:00.000Z
                      estimatedDurationMinutes: 15
                      condition:
                        value: Teploty
                        code: subfebrilia
                        reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
                    - id: x321ab4
                      name: Ferdinand Trety
                      patientId: 780907-ferdinand-tre
                      waitingSince: 2038-12-24T10:25:00.000Z
                      estimatedStart: 2038-12-24T10:50:00.000Z
                      estimatedDurationMinutes: 25
                      condition:
                        value: Nevoľnosť
                        code: nausea
                        reference: https://zdravoteka.sk/priznaky/nevolnost/
                    predefinedConditions:
                    - value: Teploty
                      code: subfebrilia
                      reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
                      typicalDurationMinutes: 20
                    - value: Nevoľnosť
                      code: nausea
                      reference: https://zdravoteka.sk/priznaky/nevolnost/
                      typicalDurationMinutes: 45
                    - value: Kontrola
                      code: followup
                      typicalDurationMinutes: 15
                    - value: Administratívny úkon
                      code: administration
                      typicalDurationMinutes: 10
                    - value: Odber krvy
                      code: blood-test
                      typicalDurationMinutes: 10
              schema:
                $ref: '#/components/schemas/Ambulance'
          description: array of the waiting list entries
        "404":
          description: Ambulance with such ID does not exists
      summary: Provides waiting list of the ambulance
      tags:
      - ambulanceDevelopers
    post:
      description: Stores new ambulance.
      operationId: createAmbulanceDetails
      parameters:
      - description: pass the ambulanceId of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ambulance'
        description: Ambulance template
        required: true
      responses:
        "200":
          description: ambulance create
        "403":
          description: Ambulance with such ID already exists
      summary: Creates new ambulance
      tags:
      - ambulanceAdmins
  /waiting-list/{ambulanceId}/entries:
    get:
      description: By using ambulanceId you get list of entries in ambulance witing
        list
      operationId: getWaitingListEntries
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                response:
                  value:
                  - id: x321ab3
                    name: Jožko Púčik
                    patientId: 460527-jozef-pucik
                    waitingSince: 2038-12-24T10:05:00.000Z
                    estimatedStart: 2038-12-24T10:35:00.000Z
                    estimatedDurationMinutes: 15
                    condition:
                      value: Teploty
                      code: subfebrilia
                      reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
                  - id: x321ab4
                    name: Ferdinand Trety
                    patientId: 780907-ferdinand-tre
                    waitingSince: 2038-12-24T10:25:00.000Z
                    estimatedStart: 2038-12-24T10:50:00.000Z
                    estimatedDurationMinutes: 25
                    condition:
                      value: Nevoľnosť
                      code: nausea
                      reference: https://zdravoteka.sk/priznaky/nevolnost/
              schema:
                items:
                  $ref: '#/components/schemas/WaitingListEntry'
                type: array
          description: value of the waiting list entries
        "404":
          description: Ambulance with such ID does not exists
      summary: Provides the ambulance waiting list
      tags:
      - ambulanceDevelopers
  /waiting-list/{ambulanceId}/entry:
    post:
      description: Use this method to store new entry into the waiting list.
      operationId: storeWaitingListEntry
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitingListEntry'
        description: Waiting list entry to store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitingListEntry'
          description: Value of the waiting list entry with re-computed estimated
            time of ambulance entry
        "400":
          description: Missing mandatory properties of input object.
        "404":
          description: Ambulance with such ID does not exists
      summary: Saves new entry into waiting list
      tags:
      - ambulanceDevelopers
  /waiting-list/{ambulanceId}/entry/{entryId}:
    delete:
      description: Use this method to delete the specific entry from the waiting list.
      operationId: deleteWaitingListEntry
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the id of the particular entry in the waiting list
        explode: false
        in: path
        name: entryId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Item deleted
        "404":
          description: Ambulance or Entry with such ID does not exists
      summary: Deletes specific entry
      tags:
      - ambulanceDevelopers
    get:
      description: By using ambulanceId and entryId you can details of particular
        entry item ambulance.
      operationId: getWaitingListEntry
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the id of the particular entry in the waiting list
        explode: false
        in: path
        name: entryId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitingListEntry'
          description: value of the waiting list entries
        "404":
          description: Ambulance or Entry with such ID does not exists
      summary: Provides details about waiting list entry
      tags:
      - ambulanceDevelopers
    post:
      description: Use this method to update content of the waiting list entry.
      operationId: updateWaitingListEntry
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the id of the particular entry in the waiting list
        explode: false
        in: path
        name: entryId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitingListEntry'
        description: Waiting list entry to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitingListEntry'
          description: value of the waiting list entry with re-computed estimated
            time of ambulance entry
        "403":
          description: Value of the entryID and the data id is mismatching. Detail
            s are provided in the response body.
        "404":
          description: Ambulance or Entry with such ID does not exists
      summary: Updates specific entry
      tags:
      - ambulanceDevelopers
  /waiting-list/{ambulanceId}/condition:
    get:
      description: By using ambulanceId you get list of predefined conditions
      operationId: getConditions
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                response:
                  value:
                  - value: Teploty
                    code: subfebrilia
                    reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
                    typicalDurationMinutes: 20
                  - value: Nevoľnosť
                    code: nausea
                    reference: https://zdravoteka.sk/priznaky/nevolnost/
                    typicalDurationMinutes: 45
                  - value: Kontrola
                    code: followup
                    typicalDurationMinutes: 15
                  - value: Administratívny úkon
                    code: administration
                    typicalDurationMinutes: 10
                  - value: Odber krvy
                    code: blood-test
                    typicalDurationMinutes: 10
              schema:
                items:
                  $ref: '#/components/schemas/Condition'
                type: array
          description: value of the predefined conditions
        "404":
          description: Ambulance with such ID does not exists
      summary: Provides the list of conditions associated with ambulance
      tags:
      - ambulanceDevelopers
    post:
      description: Use this method to associate new predefined condition with the
        ambulance
      operationId: storeCondition
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Condition'
        description: Associated condition
        required: true
      responses:
        "200":
          description: Value was stored
        "400":
          description: Missing mandatory properties of input object.
        "404":
          description: Ambulance with such ID does not exists
      summary: Saves new predefined condition
      tags:
      - ambulanceAdmins
  /waiting-list/{ambulanceId}/condition/{conditionCode}:
    delete:
      description: Use this method to delete the specific condition from the list
        of associated predefined conditions for this ambulance
      operationId: deleteCondition
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the code of the particular condition
        explode: false
        in: path
        name: conditionCode
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Item deleted
        "404":
          description: Ambulance with such ID or Condition with such code does not
            exists
      summary: Deletes specific associated condition
      tags:
      - ambulanceAdmins
    get:
      description: By using ambulanceId and conditionCode you can get details of associated
        predefined condition.
      operationId: getCondition
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the code of the particular condition
        explode: false
        in: path
        name: conditionCode
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
          description: value of the requested condition
        "404":
          description: Ambulance with such ID or Condition with such code does not
            exists
      summary: Provides details about given condition
      tags:
      - ambulanceDevelopers
    post:
      description: Use this method to update content of the associated predefined
        condition.
      operationId: updateCondition
      parameters:
      - description: pass the id of the particular ambulance
        explode: false
        in: path
        name: ambulanceId
        required: true
        schema:
          type: string
        style: simple
      - description: pass the code of the particular condition
        explode: false
        in: path
        name: conditionCode
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Condition'
        description: Condition to update
        required: true
      responses:
        "200":
          description: Value updated
        "403":
          description: Value of the conditionCode and the code in data are mismatching.
            Details are provided in the response body.
        "404":
          description: Ambulance with such
      summary: Updates specific condition
      tags:
      - ambulanceDevelopers
components:
  schemas:
    Ambulance:
      example:
        roomNumber: 356 - 3.posch
        waitingList:
        - estimatedStart: 2038-12-24T10:35:00Z
          waitingSince: 2038-12-24T10:05:00Z
          condition:
            reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
            code: subfebrilia
            typicalDurationMinutes: 20
            value: Teploty
          patientId: 460527-jozef-pucik
          estimatedDurationMinutes: 15
          name: Jožko Púčik
          id: x321ab3
        - estimatedStart: 2038-12-24T10:35:00Z
          waitingSince: 2038-12-24T10:05:00Z
          condition:
            reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
            code: subfebrilia
            typicalDurationMinutes: 20
            value: Teploty
          patientId: 460527-jozef-pucik
          estimatedDurationMinutes: 15
          name: Jožko Púčik
          id: x321ab3
        name: Zubná ambulancia Dr. Warenová
        id: dentist-warenova
        predefinedConditions:
        - reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
          code: subfebrilia
          typicalDurationMinutes: 20
          value: Teploty
        - reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
          code: subfebrilia
          typicalDurationMinutes: 20
          value: Teploty
      properties:
        id:
          description: Unique identifier of the ambulance
          example: dentist-warenova
          type: string
        name:
          description: Human readable display name of the ambulance
          example: Zubná ambulancia Dr. Warenová
          type: string
        roomNumber:
          example: 356 - 3.posch
          type: string
        waitingList:
          items:
            $ref: '#/components/schemas/WaitingListEntry'
          type: array
        predefinedConditions:
          items:
            $ref: '#/components/schemas/Condition'
          type: array
      required:
      - id
      - name
      - roomNumber
      type: object
    WaitingListEntry:
      example:
        estimatedStart: 2038-12-24T10:35:00Z
        waitingSince: 2038-12-24T10:05:00Z
        condition:
          reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
          code: subfebrilia
          typicalDurationMinutes: 20
          value: Teploty
        patientId: 460527-jozef-pucik
        estimatedDurationMinutes: 15
        name: Jožko Púčik
        id: x321ab3
      properties:
        id:
          description: Unique id of the entry in this waiting list
          example: x321ab3
          type: string
        name:
          description: Name of patient in waiting list
          example: Jožko Púčik
          type: string
        patientId:
          description: Unique identifier of the patient known to Web-In-Cloud system
          example: 460527-jozef-pucik
          type: string
        waitingSince:
          description: Timestamp since when the patient entered the waiting list
          example: 2038-12-24T10:05:00Z
          format: date-time
          type: string
        estimatedStart:
          description: Estimated time of entering ambulance. Ignored on post.
          example: 2038-12-24T10:35:00Z
          format: date-time
          type: string
        estimatedDurationMinutes:
          description: Estimated duration of ambulance visit. If not provided then
            it will be computed based on condition and ambulance settings
          example: 15
          format: int32
          type: integer
        condition:
          $ref: '#/components/schemas/Condition'
      required:
      - estimatedDurationMinutes
      - id
      - patientId
      - waitingSince
      type: object
    Condition:
      description: "Describes disease, symptoms, or other reasons of patient visit"
      example:
        reference: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
        code: subfebrilia
        typicalDurationMinutes: 20
        value: Teploty
      properties:
        value:
          example: Teploty
          type: string
        code:
          example: subfebrilia
          type: string
        reference:
          description: Link to encyclopedical explanation of the patient's condition
          example: https://zdravoteka.sk/priznaky/zvysena-telesna-teplota/
          format: url
          type: string
        typicalDurationMinutes:
          example: 20
          format: int32
          type: integer
      required:
      - value