# models/model_ambulance.go
/*
 * Waiting List Api
 *
 * Ambulance Waiting List management for Web-In-Cloud system
 *
 * API version: 1.0.0
 * Contact: aa@bb.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type Ambulance struct {

	// Unique identifier of the ambulance
	Id string `json:"id" bson:"Id"`

	// Human readable display name of the ambulance
	Name string `json:"name" bson:"Name,omitempty"`

	RoomNumber string `json:"roomNumber" bson:"RoomNumber,omitempty"`

	WaitingList []WaitingListEntry `json:"waitingList,omitempty" bson:"WaitingList,omitempty"`

	PredefinedConditions []Condition `json:"predefinedConditions,omitempty" bson:"PredefinedConditions,omitempty"`
}

func (ambulance *Ambulance) ContainsWaitingListEntry(waitingListEntry WaitingListEntry) bool {
	for _, entry := range ambulance.WaitingList {
		if entry.PatientId == waitingListEntry.PatientId {
			return true
		}
	}

	return false
}