# models/model_condition.go
/*
 * Waiting List Api
 *
 * Ambulance Waiting List management for Web-In-Cloud system
 *
 * API version: 1.0.0
 * Contact: aa@bb.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

// Condition - Describes disease, symptoms, or other reasons of patient visit
type Condition struct {
	Value string `json:"value" bson:"Value"`

	Code string `json:"code,omitempty" bson:"Code,omitempty"`

	// Link to encyclopedical explanation of the patient's condition
	Reference string `json:"reference,omitempty" bson:"Reference,omitempty"`

	TypicalDurationMinutes int32 `json:"typicalDurationMinutes,omitempty" bson:"TypicalDurationMinutes,omitempty"`
}