# models/model_waiting_list_entry.go
/*
 * Waiting List Api
 *
 * Ambulance Waiting List management for Web-In-Cloud system
 *
 * API version: 1.0.0
 * Contact: aa@bb.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

import (
	"time"
)

type WaitingListEntry struct {

	// Unique id of the entry in this waiting list
	Id string `json:"id" bson:"Id"`

	// Name of patient in waiting list
	Name string `json:"name,omitempty" bson:"Name,omitempty"`

	// Unique identifier of the patient known to Web-In-Cloud system
	PatientId string `json:"patientId" bson:"PatientId"`

	// Timestamp since when the patient entered the waiting list
	WaitingSince time.Time `json:"waitingSince" bson:"WaitingSince"`

	// Estimated time of entering ambulance. Ignored on post.
	EstimatedStart time.Time `json:"estimatedStart,omitempty" bson:"EstimatedStart,omitempty"`

	// Estimated duration of ambulance visit. If not provided then it will be computed based on condition and ambulance settings
	EstimatedDurationMinutes int32 `json:"estimatedDurationMinutes" bson:"EstimatedDurationMinutes"`

	Condition Condition `json:"condition,omitempty" bson:"Condition,omitempty"`
}