# rest-api/api_ambulance_admins.go
/*
 * Waiting List Api
 *
 * Ambulance Waiting List management for Web-In-Cloud system
 *
 * API version: 1.0.0
 * Contact: aa@bb.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package restapi

import (
	dbservice "ambulance-webapi/db-service"
	"ambulance-webapi/models"
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
)

// CreateAmbulanceDetails - Creates new ambulance
func CreateAmbulanceDetails(c *gin.Context) {
	var ambulance models.Ambulance

	if err := c.ShouldBindJSON(&ambulance); err != nil {
		log.Println(err)
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	err := dbservice.CreateAmbulance(&ambulance)
	if err != nil {
		log.Println(err)
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	log.Println("CreateAmbulanceDetails successfully created ambulance with id: ", ambulance.Id)
	c.JSON(http.StatusCreated, gin.H{"ambulanceId": ambulance.Id})
}

// DeleteCondition - Deletes specific associated condition
func DeleteCondition(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// StoreCondition - Saves new predefined condition
func StoreCondition(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}